refactor preprocess transforms
