+++ pennylane/devices/default_qubit.py
@@ -15,6 +15,7 @@
 This module contains the next generation successor to default qubit
 """
 
+from dataclasses import replace
 from functools import partial
 from numbers import Number
 from typing import Union, Callable, Tuple, Optional, Sequence
@@ -28,15 +29,16 @@ from pennylane.transforms import convert_to_numpy_parameters
 from pennylane.transforms.core import TransformProgram
 
 from . import Device
+from .preprocess import (
+    decompose,
+    validate_measurements,
+    validate_multiprocessing_workers,
+    validate_device_wires,
+    warn_about_trainable_observables,
+)
 from .execution_config import ExecutionConfig, DefaultExecutionConfig
 from .qubit.simulate import simulate, get_final_state, measure_final_state
 from .qubit.sampling import get_num_shots_and_executions
-from .qubit.preprocess import (
-    preprocess,
-    validate_and_expand_adjoint,
-    validate_multiprocessing_workers,
-    validate_device_wires,
-)
 from .qubit.adjoint_jacobian import adjoint_jacobian, adjoint_vjp, adjoint_jvp
 
 Result_or_ResultBatch = Union[Result, ResultBatch]
@@ -44,6 +46,53 @@ QuantumTapeBatch = Sequence[QuantumTape]
 QuantumTape_or_Batch = Union[QuantumTape, QuantumTapeBatch]
 # always a function from a resultbatch to either a result or a result batch
 PostprocessingFn = Callable[[ResultBatch], Result_or_ResultBatch]
+
+
+observables = {
+    "PauliX",
+    "PauliY",
+    "PauliZ",
+    "Hadamard",
+    "Hermitian",
+    "Identity",
+    "Projector",
+    "SparseHamiltonian",
+    "Hamiltonian",
+    "Sum",
+    "SProd",
+    "Prod",
+    "Exp",
+    "Evolution",
+}
+
+
+def observable_stopping_coindition(obs: qml.operation.Operator) -> bool:
+    """Specifies whether or not an observable is accepted by DefaultQubit."""
+    return obs.name in observables
+
+
+def adjoint_observable_stopping_condition(obs: qml.operation.Operator) -> bool:
+    """Specifies whether or not an observable is compatible with adjoint differentiation on DefaultQubit."""
+    return obs.has_matrix
+
+
+def stopping_condition(op: qml.operation.Operator) -> bool:
+    """Specify whether or not an Operator object is supported by the device."""
+    if op.name == "QFT" and len(op.wires) >= 6:
+        return False
+    if op.name == "GroverOperator" and len(op.wires) >= 13:
+        return False
+    if op.name == "Snapshot":
+        return True
+    if op.__class__.__name__ == "Pow" and qml.operation.is_trainable(op):
+        return False
+
+    return op.has_matrix
+
+
+def adjoint_stopping_condition(op: qml.operation.Operator) -> bool:
+    """Specify whether or not an Oeprator is supported by adjoint differentiation."""
+    return op.num_params == 0 or op.num_params == 1 and op.has_generator
 
 
 class DefaultQubit(Device):
@@ -249,7 +298,11 @@ class DefaultQubit(Device):
             if circuit is None:
                 return True
 
-            return isinstance(validate_and_expand_adjoint(circuit)[0][0], QuantumScript)
+            try:
+                decompose(circuit, stopping_condition=adjoint_stopping_condition)
+            except (qml.operation.DecompositionUndefinedError, qml.DeviceError):
+                return False
+            return True
 
         return False
 
@@ -276,6 +329,12 @@ class DefaultQubit(Device):
 
         """
         transform_program = TransformProgram()
+
+        transform_program.add_transform(decompose, stopping_condition=stopping_condition)
+        transform_program.add_transform(
+            validate_measurements, observable_stopping_condition=observable_stopping_coindition
+        )
+
         # Validate device wires
         transform_program.add_transform(validate_device_wires, self)
 
@@ -283,10 +342,40 @@ class DefaultQubit(Device):
         max_workers = execution_config.device_options.get("max_workers", self._max_workers)
         transform_program.add_transform(validate_multiprocessing_workers, max_workers, self)
 
-        # General preprocessing (Validate measurement, expand, adjoint expand, broadcast expand)
-        transform_program_preprocess, config = preprocess(execution_config=execution_config)
-        transform_program = transform_program + transform_program_preprocess
-        return transform_program, config
+        if execution_config.gradient_method == "adjoint":
+            transform_program.add_transform(
+                decompose, stopping_condition=adjoint_stopping_condition
+            )
+            transform_program.add_transform(
+                validate_measurements, adjoint_observable_stopping_condition
+            )
+            transform_program.add_transform(qml.transforms.broadcast_expand)
+            transform_program.add_transform(warn_about_trainable_observables)
+
+        return transform_program, self._setup_execution_config(execution_config)
+
+    def _setup_execution_config(self, execution_config: ExecutionConfig) -> ExecutionConfig:
+        """This is a private helper for ``preprocess`` that sets up the execution config.
+
+        Args:
+            execution_config (ExecutionConfig)
+
+        Returns:
+            ExecutionConfig: a preprocessed execution config
+
+        """
+        updated_values = {}
+        if execution_config.gradient_method == "best":
+            updated_values["gradient_method"] = "backprop"
+        if execution_config.use_device_gradient is None:
+            updated_values["use_device_gradient"] = execution_config.gradient_method in {
+                "best",
+                "adjoint",
+                "backprop",
+            }
+        if execution_config.grad_on_execution is None:
+            updated_values["grad_on_execution"] = execution_config.gradient_method == "adjoint"
+        return replace(execution_config, **updated_values)
 
     def execute(
         self,
